Use combined-stream2 library
combined-stream2 is compatible with Streams v2 API, and therefore plays
nicer with the libraries we use in our services, for example
event-stream.

This fixes an issue we have currently on image-maker, where event-stream
writes too much data to the zip library, before that is able to write it
to the response, throwing "CombinedStream#maxDataSize max data size exceeded".