Add DEFLATE_END block to the end of each entry.
This allows all deflate parts to be the same, and have the same
format as gzip-stream.

This was supposed to be part of 0.1.0 but was accidentally not
included.

Tests were passing because of a failure in the test framework,
it wasn't actually running the expect() checks. Tests are fixed
and now automatically generate all the required files instead
of pre-generating them as was done until now.